move DataSet stuff except Dao into package
