+++ api/src/main/java/org/pmiops/workbench/api/CohortsController.java
@@ -58,7 +58,7 @@ public class CohortsController implements CohortsApiDelegate {
    * Converter function from backend representation (used with Hibernate) to client representation
    * (generated by Swagger).
    */
-  static final Function<org.pmiops.workbench.db.model.Cohort, Cohort> TO_CLIENT_COHORT =
+  public static final Function<org.pmiops.workbench.db.model.Cohort, Cohort> TO_CLIENT_COHORT =
       new Function<org.pmiops.workbench.db.model.Cohort, Cohort>() {
         @Override
         public Cohort apply(org.pmiops.workbench.db.model.Cohort cohort) {
@@ -92,7 +92,7 @@ public class CohortsController implements CohortsApiDelegate {
   private final UserRecentResourceService userRecentResourceService;
 
   @Autowired
-  CohortsController(
+  public CohortsController(
       WorkspaceService workspaceService,
       CohortDao cohortDao,
       CdrVersionDao cdrVersionDao,
