+++ api/src/main/java/org/pmiops/workbench/dataset/DataSetServiceImpl.java
@@ -1,4 +1,4 @@
-package org.pmiops.workbench.db.dao;
+package org.pmiops.workbench.dataset;
 
 import com.google.cloud.bigquery.QueryJobConfiguration;
 import com.google.cloud.bigquery.QueryParameterValue;
@@ -24,6 +24,9 @@ import org.pmiops.workbench.cohortbuilder.ParticipantCriteria;
 import org.pmiops.workbench.config.CdrBigQuerySchemaConfig;
 import org.pmiops.workbench.config.CdrBigQuerySchemaConfigService;
 import org.pmiops.workbench.config.WorkbenchConfig;
+import org.pmiops.workbench.db.dao.CohortDao;
+import org.pmiops.workbench.db.dao.ConceptSetDao;
+import org.pmiops.workbench.db.dao.DataSetDao;
 import org.pmiops.workbench.db.model.Cohort;
 import org.pmiops.workbench.db.model.CommonStorageEnums;
 import org.pmiops.workbench.db.model.ConceptSet;
@@ -98,11 +101,14 @@ public class DataSetServiceImpl implements DataSetService {
   private CdrBigQuerySchemaConfigService cdrBigQuerySchemaConfigService;
   private Provider<WorkbenchConfig> configProvider;
 
-  @Autowired DataSetDao dataSetDao;
+  @Autowired
+  DataSetDao dataSetDao;
 
-  @Autowired ConceptSetDao conceptSetDao;
+  @Autowired
+  ConceptSetDao conceptSetDao;
 
-  @Autowired CohortDao cohortDao;
+  @Autowired
+  CohortDao cohortDao;
 
   @Autowired
   @VisibleForTesting
