+++ api/src/test/java/org/pmiops/workbench/dataset/DataSetControllerTest.java
@@ -1,4 +1,4 @@
-package org.pmiops.workbench.api;
+package org.pmiops.workbench.dataset;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.mockito.Mockito.*;
@@ -32,6 +32,9 @@ import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.invocation.InvocationOnMock;
+import org.pmiops.workbench.api.BigQueryService;
+import org.pmiops.workbench.api.CohortsController;
+import org.pmiops.workbench.api.ConceptSetsController;
 import org.pmiops.workbench.billing.BillingProjectBufferService;
 import org.pmiops.workbench.cdr.CdrVersionService;
 import org.pmiops.workbench.cdr.ConceptBigQueryService;
@@ -46,13 +49,14 @@ import org.pmiops.workbench.conceptset.ConceptSetService;
 import org.pmiops.workbench.config.CdrBigQuerySchemaConfig;
 import org.pmiops.workbench.config.CdrBigQuerySchemaConfigService;
 import org.pmiops.workbench.config.WorkbenchConfig;
+import org.pmiops.workbench.dataset.DataSetController;
 import org.pmiops.workbench.db.dao.CdrVersionDao;
 import org.pmiops.workbench.db.dao.CohortDao;
 import org.pmiops.workbench.db.dao.CohortReviewDao;
 import org.pmiops.workbench.db.dao.ConceptSetDao;
 import org.pmiops.workbench.db.dao.DataSetDao;
-import org.pmiops.workbench.db.dao.DataSetService;
-import org.pmiops.workbench.db.dao.DataSetServiceImpl;
+import org.pmiops.workbench.dataset.DataSetService;
+import org.pmiops.workbench.dataset.DataSetServiceImpl;
 import org.pmiops.workbench.db.dao.UserDao;
 import org.pmiops.workbench.db.dao.UserRecentResourceService;
 import org.pmiops.workbench.db.dao.UserService;
@@ -144,7 +148,8 @@ public class DataSetControllerTest {
 
   @Autowired BillingProjectBufferService billingProjectBufferService;
 
-  @Autowired BigQueryService bigQueryService;
+  @Autowired
+  BigQueryService bigQueryService;
 
   @Autowired CdrBigQuerySchemaConfigService cdrBigQuerySchemaConfigService;
